requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
asyncer==0.0.7
attrs==25.3.0
autogen-agentchat==0.4.5
autogen-core==0.4.5
autogen-ext==0.4.5
av==14.4.0
azure-ai-agents==1.1.0b3
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b11
azure-common==1.1.28
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-search-documents==11.6.0b12
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chainlit==2.5.5
chardet==5.2.0
charset-normalizer==3.4.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==45.0.4
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
dnspython==2.7.0
dotenv==0.9.9
durationpy==0.10
executing==2.2.0
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
grpcio==1.73.1
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kubernetes==33.1.0
Lazify==0.4.0
lazy-object-proxy==1.11.0
literalai==0.1.201
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.10.1
mdurl==0.1.2
mistralai==0.4.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.40.14
opentelemetry-instrumentation-anthropic==0.40.14
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-bedrock==0.40.14
opentelemetry-instrumentation-chromadb==0.40.14
opentelemetry-instrumentation-cohere==0.40.14
opentelemetry-instrumentation-crewai==0.40.14
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-google-generativeai==0.40.14
opentelemetry-instrumentation-groq==0.40.14
opentelemetry-instrumentation-haystack==0.40.14
opentelemetry-instrumentation-lancedb==0.40.14
opentelemetry-instrumentation-langchain==0.40.14
opentelemetry-instrumentation-llamaindex==0.40.14
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.40.14
opentelemetry-instrumentation-mcp==0.40.14
opentelemetry-instrumentation-milvus==0.40.14
opentelemetry-instrumentation-mistralai==0.40.14
opentelemetry-instrumentation-ollama==0.40.14
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-pinecone==0.40.14
opentelemetry-instrumentation-qdrant==0.40.14
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.40.14
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.40.14
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.40.14
opentelemetry-instrumentation-transformers==0.40.14
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.40.14
opentelemetry-instrumentation-watsonx==0.40.14
opentelemetry-instrumentation-weaviate==0.40.14
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.3.0
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pdf2image==1.17.0
pillow==11.0.0
platformdirs==4.3.8
posthog==3.25.0
prance==25.4.8.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.9.0
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-multipart==0.0.18
python-slugify==8.0.4
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scipy==1.16.0
semantic-kernel==1.34.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.41.3
streamlit==1.46.1
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.40.14
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uptrace==1.34.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
